Refactor cryptocurrency system to store pricing and data in database
- Update price and market data for all tracked cryptocurrencies periodically from external APIs
- Replace Double with BigDecimal for precise handling of large and fractional values
- Refactor methods for buying and selling cryptocurrency to align with new database-based pricing model
- Add 60-second interval refresh for cryptocurrency prices on frontend
- Implement backend scheduling tasks to automatically fetch and update cryptocurrency information in the database at defined intervals

+++ Trading-Simulator/backend/src/main/java/com/example/backend/transaction/BuyAssetRequest.java
@@ -7,6 +7,8 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class BuyAssetRequest {
@@ -19,5 +21,5 @@ public class BuyAssetRequest {
     @NotNull(message = "Amount in USD is required")
     @Min(value = 0, message = "Amount must be positive")
     @JsonProperty("amountInUSD")
-    private Double amountInUSD;
+    private BigDecimal amountInUSD;
 }
