+++ Trading-Simulator/backend/src/main/java/com/example/backend/transaction/TransactionHistoryDTO.java
@@ -2,6 +2,7 @@ package com.example.backend.transaction;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -12,8 +13,8 @@ import java.time.LocalDateTime;
 public class TransactionHistoryDTO {
     private Integer transactionid;
     private String transactionType;
-    private Double amount;
-    private Double rate;
+    private BigDecimal amount;
+    private BigDecimal rate;
     private LocalDateTime timestamp;
     private String currencyName;
     private String portfolioName;
